# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOSHI Wallet API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  name: string;
  address: string;
  chain: string;
  isActive: boolean;
  totalBalanceUsd: number;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface WalletInput {
  name: string;
  address: string;
  chain: string;
}

export interface WalletUpdate {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface ChartPoint {
  label: string;
  value: number;
}

export interface PeriodGain {
  period: string;
  gainUsd: number;
  gainPct: number;
}

export interface PortfolioSummary {
  totalBalanceUsd: number;
  gainUsd: number;
  gainPct: number;
  activeAssets: number;
  diversificationPct: number;
  lastUpdated?: string;
  chartData: ChartPoint[];
  gainsByPeriod: PeriodGain[];
}

export interface Token {
  id: number;
  symbol: string;
  name: string;
  chain: string;
  /** @nullable */
  logoUrl?: string | null;
  balance: number;
  balanceUsd: number;
  priceUsd: number;
  change24h: number;
  isFavorite?: boolean;
  isEnabled: boolean;
  /** @nullable */
  contractAddress?: string | null;
}

export interface TokenInfo {
  symbol: string;
  name: string;
  chain: string;
  /** @nullable */
  logoUrl?: string | null;
  priceUsd: number;
  rank: number;
  /** @nullable */
  contractAddress?: string | null;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  send: 'send',
  receive: 'receive',
  swap: 'swap',
  buy: 'buy',
  sell: 'sell',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  tokenSymbol: string;
  /** @nullable */
  tokenLogoUrl?: string | null;
  amount: number;
  amountUsd: number;
  chain: string;
  /** @nullable */
  fromAddress?: string | null;
  /** @nullable */
  toAddress?: string | null;
  /** @nullable */
  txHash?: string | null;
  status: TransactionStatus;
  /** @nullable */
  gasUsd?: number | null;
  createdAt: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  send: 'send',
  receive: 'receive',
  swap: 'swap',
  buy: 'buy',
  sell: 'sell',
} as const;

export interface TransactionInput {
  type: TransactionInputType;
  tokenSymbol: string;
  amount: number;
  amountUsd: number;
  chain: string;
  /** @nullable */
  fromAddress?: string | null;
  /** @nullable */
  toAddress?: string | null;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  gasUsd?: number | null;
}

export interface MarketOverview {
  totalMarketCapUsd: number;
  btcDominancePct: number;
  ethDominancePct: number;
  volume24hUsd: number;
  change24hPct: number;
  fearGreedIndex: number;
  fearGreedLabel: string;
}

export interface MarketToken {
  symbol: string;
  name: string;
  chain: string;
  /** @nullable */
  logoUrl?: string | null;
  priceUsd: number;
  change24h: number;
  volume24hUsd: number;
  rank: number;
  /** @nullable */
  contractAddress?: string | null;
}

export type NewsArticleSentiment = typeof NewsArticleSentiment[keyof typeof NewsArticleSentiment];


export const NewsArticleSentiment = {
  bullish: 'bullish',
  bearish: 'bearish',
  neutral: 'neutral',
} as const;

export interface NewsArticle {
  id: number;
  title: string;
  source: string;
  publishedAt: string;
  sentiment: NewsArticleSentiment;
  /** @nullable */
  imageUrl?: string | null;
  url: string;
  /** @nullable */
  summary?: string | null;
}

export interface Nft {
  id: number;
  name: string;
  collection: string;
  chain: string;
  /** @nullable */
  imageUrl?: string | null;
  tokenId: string;
  /** @nullable */
  contractAddress?: string | null;
  /** @nullable */
  floorPriceUsd?: number | null;
  /** @nullable */
  description?: string | null;
}

export interface AffiliateStats {
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
  volumeSwapUsd: number;
  commissionUsd: number;
  tokensEarned: number;
  tokensClaimed: number;
}

export type SettingsTheme = typeof SettingsTheme[keyof typeof SettingsTheme];


export const SettingsTheme = {
  dark: 'dark',
  light: 'light',
  system: 'system',
} as const;

export interface Settings {
  id: number;
  currency: string;
  language: string;
  theme: SettingsTheme;
  hideBalances: boolean;
  biometricsEnabled: boolean;
  autoLockMinutes: number;
  notifyPriceAlerts: boolean;
  notifyTransactions: boolean;
  notifySecurity: boolean;
}

export type SettingsUpdateTheme = typeof SettingsUpdateTheme[keyof typeof SettingsUpdateTheme];


export const SettingsUpdateTheme = {
  dark: 'dark',
  light: 'light',
  system: 'system',
} as const;

export interface SettingsUpdate {
  currency?: string;
  language?: string;
  theme?: SettingsUpdateTheme;
  hideBalances?: boolean;
  biometricsEnabled?: boolean;
  autoLockMinutes?: number;
  notifyPriceAlerts?: boolean;
  notifyTransactions?: boolean;
  notifySecurity?: boolean;
}