# lib/api-zod/src/generated/types/affiliateStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOSHI Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface AffiliateStats {
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
  volumeSwapUsd: number;
  commissionUsd: number;
  tokensEarned: number;
  tokensClaimed: number;
}