# lib/api-zod/src/generated/types/marketOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOSHI Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketOverview {
  totalMarketCapUsd: number;
  btcDominancePct: number;
  ethDominancePct: number;
  volume24hUsd: number;
  change24hPct: number;
  fearGreedIndex: number;
  fearGreedLabel: string;
}