# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOSHI Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticleSentiment } from './newsArticleSentiment';

export interface NewsArticle {
  id: number;
  title: string;
  source: string;
  publishedAt: string;
  sentiment: NewsArticleSentiment;
  /** @nullable */
  imageUrl?: string | null;
  url: string;
  /** @nullable */
  summary?: string | null;
}