# lib/api-zod/src/generated/types/periodGain.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOSHI Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface PeriodGain {
  period: string;
  gainUsd: number;
  gainPct: number;
}